Adding a proposed mapping between SQL database types and Rust.
This mapping assumes that writes are done via changesets and is
thus only one-directional. A `Mapping<T>` is generic for any struct
that wants to be mappable from SQL.

A user struct `Foo` would then need to implement `From<SqlType<'a> for Foo`
and a new mapping type which is `impl<'a> Mapping<'a, Foo> for FooMap {}` which
enables mapping via `FooMap::from_sql(&row)` where `row` is returned from
an sqlite query.

Ideally most of this code will be generated for the user, making manual
implementations for `From<?>` redundant.

`db_tool` adds some simple utilities to work with the database outside
of the example code and should be removed soon.